requirements.txt
requests==2.22.0
discord.py==1.7.3
RPi.GPIO==0.7.1
python-dotenv==0.20.0